refactor retry
